+++ CHANGES.md
@@ -78,6 +78,9 @@
  - The `coq-lsp.pp_type` VSCode client option now takes effect
    immediately, no more need to restart the server to get different
    goal display formats (@ejgallego, #675)
+ - new public VSCode extension API so other extensions can perform
+   actions when the user request the goals (@ejgallego, @bhaktishh,
+   #672, fixes #538)
 
 # coq-lsp 0.1.8.1: Spring fix
 -----------------------------